Home

BioGenerics

BioGenerics.Exceptions.MissingFieldException

BioGenerics.IO.AbstractFormattedIO

BioGenerics.IO.AbstractReader

BioGenerics.IO.AbstractWriter

BioGenerics.IO.stream

BioGenerics.IO.tryread!

BioGenerics.Testing.intempdir

BioGenerics.Testing.random_aa

BioGenerics.Testing.random_array

BioGenerics.Testing.random_dna

BioGenerics.Testing.random_interval

BioGenerics.Testing.random_rna

BioGenerics.Testing.random_seq

BioGenerics.IO.AbstractFormattedIO — Type.

Abstract formatted input/output type.

source

BioGenerics.IO.AbstractReader — Type.

Abstract data reader type.

See subtypes(AbstractReader) for all available data readers.

source

BioGenerics.IO.AbstractWriter — Type.

Abstract data writer type.

See subtypes(AbstractWriter) for all available data writers.

source

BioGenerics.IO.stream — Function.

stream(io::AbstractFormattedIO)

Return the underlying IO object; subtypes of AbstractFormattedIO must implement this method.

source

BioGenerics.IO.tryread! — Method.

tryread!(reader::AbstractReader, output)

Try to read the next element into output from reader.

If the result could not be read, then nothing will be returned instead.

source

BioGenerics.Exceptions.MissingFieldException — Type.

MissingFieldException <: Exception

An exception type thrown when a missing field of a record is accessed.

source

BioGenerics.Testing.intempdir — Function.

intempdir(fn::Function, parent = tempdir())

Execute some function fn in a temporary directory. After the function is executed, the directory is cleaned by doing the equivalent of rm -r.

source

BioGenerics.Testing.random_aa — Method.

random_aa(n, probs = [0.24, 0.24, 0.24, 0.24, 0.04])

Create a random amino acid sequence of length n, by sampling the possible amino acid characters.

source

BioGenerics.Testing.random_array — Method.

random_array(n::Integer, elements, probs)

Create a random array of length n, composed of possible elements, which are selected according to their prob.

source

BioGenerics.Testing.random_dna — Function.

random_dna(n, probs = [0.24, 0.24, 0.24, 0.24, 0.04])

Create a random DNA sequence of length n, by sampling the nucleotides A, C, G, T, and N, according to their probability in probs.

source

BioGenerics.Testing.random_interval — Method.

random_interval(minstart, maxstop)

Create a random interval between a minimum starting point, and a maximum stop point.

source

BioGenerics.Testing.random_rna — Function.

random_rna(n, probs = [0.24, 0.24, 0.24, 0.24, 0.04])

Create a random RNA sequence of length n, by sampling the nucleotides A, C, G, U, and N, according to their probability in probs.

source

BioGenerics.Testing.random_seq — Method.

random_seq(n::Integer, nts, probs)

Create a random sequence of length n, composed of nts, according to their prob.

source